# backend/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: NILIN API
  version: '1.0'
  description: Home Service Marketplace Backend API
  contact:
    name: NILIN Support
    email: support@nilin.app
servers:
  - url: /api
    description: API v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /automation/welcome:
    post:
      tags: [Admin]
      summary: Trigger welcome automation
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                email:
                  type: string
                  format: email
                name:
                  type: string
                source:
                  type: string
      responses:
        '200':
          description: Welcome automation triggered
  /automation/winback:
    post:
      tags: [Admin]
      summary: Trigger win-back campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
                reason:
                  type: string
                offerCode:
                  type: string
      responses:
        '200':
          description: Win-back automation triggered
  /automation/status:
    get:
      tags: [Admin]
      summary: Get automation status
      parameters:
        - in: query
          name: automationType
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Automation status
  /automation/review-request:
    post:
      tags: [Admin]
      summary: Trigger review request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookingId:
                  type: string
                customerId:
                  type: string
                delay:
                  type: integer
                  default: 24
      responses:
        '200':
          description: Review request scheduled

  # ========================================
  # Batch Operations
  # ========================================
...